python/requirements.txt
opencv-python>=4.8.0
mediapipe==0.10.32
Flask>=3.0.0
Flask-SocketIO>=5.3.0
Flask-Cors>=4.0.0
eventlet>=0.33.3